HexViewer: seprated to smaller modules
