+++ gui/widgets/hex_viewer/__init__.py
@@ -1,0 +1,7 @@
+from .hex_area import HexArea
+from .hex_viewer import HexViewer
+
+__all__ = [
+    "HexArea",
+    "HexViewer",
+]
